extract PathView into separate files
# Conflicts:
#	src/path/path.cpp
