+++ src/path/CMakeLists.txt
@@ -7,10 +7,12 @@ target_sources(libommpfritt PRIVATE
   graph.h
   lib2geomadapter.cpp
   lib2geomadapter.h
-  pathvector.h
-  pathvector.cpp
   pathpoint.cpp
   pathpoint.h
   path.cpp
   path.h
+  pathvector.cpp
+  pathvector.h
+  pathview.cpp
+  pathview.h
 )
